# perf(parser): Overhaul how parsing is done