--- a/benches/tokenize.rs
+++ b/benches/tokenize.rs
@@ -15,7 +15,8 @@ fn bench_tokenize(c: &mut Criterion) {
             b.iter(|| parser.parse_str(sample).last());
         });
         group.bench_with_input(BenchmarkId::new("words", name), &len, |b, _| {
-            let symbol = typos::tokens::Identifier::new_unchecked(sample, 0);
+            let symbol =
+                typos::tokens::Identifier::new_unchecked(sample, typos::tokens::Case::None, 0);
             b.iter(|| symbol.split().last());
         });
         group.bench_with_input(