From the user-facing perspective, we are deferring to the Unicode XID spec for the definition of an identifier, rather than having the user specify it.  We are also offering an ASCII fast path via `--no-unicode` / `default.unicode = false`.

Internally, we are switching from `regex` to `nom` for parsing.  I'm hoping this will give us some more flexibility in the parser, like email detection if we want it.

Before
```
parse_str/unicode/empty
                        time:   [32.826 ns 33.207 ns 33.590 ns]
parse_str/unicode/no_tokens
                        time:   [58.848 ns 59.315 ns 59.890 ns]
parse_str/unicode/single_token
                        time:   [95.329 ns 95.700 ns 96.133 ns]
parse_str/unicode/sherlock
                        time:   [5.1198 us 5.2072 us 5.3122 us]
parse_str/unicode/code
                        time:   [3.4724 us 3.5158 us 3.5681 us]
parse_str/unicode/corpus
                        time:   [4.4684 ms 4.5293 ms 4.5967 ms]

parse_bytes/unicode/empty
                        time:   [27.758 ns 27.853 ns 27.953 ns]
parse_bytes/unicode/no_tokens
                        time:   [56.718 ns 57.923 ns 59.160 ns]
parse_bytes/unicode/single_token
                        time:   [94.926 ns 95.558 ns 96.275 ns]
parse_bytes/unicode/sherlock
                        time:   [5.0579 us 5.1229 us 5.2037 us]
parse_bytes/unicode/code
                        time:   [3.6243 us 3.7211 us 3.8205 us]
parse_bytes/unicode/corpus
                        time:   [4.8451 ms 4.8972 ms 4.9549 ms]

split/words/empty    time:   [16.642 ns 16.878 ns 17.152 ns]
split/words/no_tokens
                        time:   [84.481 ns 85.451 ns 86.844 ns]
split/words/single_token
                        time:   [41.610 ns 41.733 ns 41.884 ns]
split/words/sherlock time:   [1.8127 us 1.8486 us 1.8865 us]
split/words/code     time:   [1.3670 us 1.3779 us 1.3898 us]
split/words/corpus   time:   [2.5282 ms 2.5582 ms 2.5912 ms]

parse_bytes+split/unicode/empty
                        time:   [75.600 ns 76.502 ns 77.457 ns]
parse_bytes+split/unicode/no_tokens
                        time:   [100.02 ns 100.56 ns 101.27 ns]
parse_bytes+split/unicode/single_token
                        time:   [215.54 ns 218.05 ns 220.68 ns]
parse_bytes+split/unicode/sherlock
                        time:   [9.6575 us 9.7123 us 9.7801 us]
parse_bytes+split/unicode/code
                        time:   [6.7650 us 6.8942 us 7.0305 us]
parse_bytes+split/unicode/corpus
                        time:   [9.8030 ms 9.8366 ms 9.8735 ms]
```

And after:
```
parse_str/unicode/empty time:   [21.279 ns 21.364 ns 21.456 ns]
parse_str/ascii/empty   time:   [14.988 ns 15.068 ns 15.150 ns]
parse_str/unicode/no_tokens
                        time:   [57.654 ns 58.639 ns 59.639 ns]
parse_str/ascii/no_tokens
                        time:   [29.964 ns 30.356 ns 30.812 ns]
parse_str/unicode/single_token
                        time:   [41.829 ns 42.053 ns 42.360 ns]
parse_str/ascii/single_token
                        time:   [29.028 ns 29.508 ns 30.006 ns]
parse_str/unicode/sherlock
                        time:   [1.7378 us 1.7623 us 1.7894 us]
parse_str/ascii/sherlock
                        time:   [1.0201 us 1.0242 us 1.0285 us]
parse_str/unicode/code  time:   [1.2219 us 1.2423 us 1.2641 us]
parse_str/ascii/code    time:   [770.91 ns 781.54 ns 793.91 ns]
parse_str/unicode/corpus
                        time:   [1.6449 ms 1.6578 ms 1.6727 ms]
parse_str/ascii/corpus  time:   [1.0435 ms 1.0616 ms 1.0818 ms]

parse_bytes/unicode/empty
                        time:   [39.988 ns 40.436 ns 40.905 ns]
parse_bytes/ascii/empty time:   [19.631 ns 19.898 ns 20.200 ns]
parse_bytes/unicode/no_tokens
                        time:   [103.35 ns 104.41 ns 105.72 ns]
parse_bytes/ascii/no_tokens
                        time:   [34.545 ns 34.760 ns 34.994 ns]
parse_bytes/unicode/single_token
                        time:   [83.197 ns 84.613 ns 86.174 ns]
parse_bytes/ascii/single_token
                        time:   [33.632 ns 33.879 ns 34.175 ns]
parse_bytes/unicode/sherlock
                        time:   [1.8156 us 1.8238 us 1.8333 us]
parse_bytes/ascii/sherlock
                        time:   [1.0493 us 1.0686 us 1.0877 us]
parse_bytes/unicode/code
                        time:   [1.2623 us 1.2744 us 1.2894 us]
parse_bytes/ascii/code  time:   [766.84 ns 770.89 ns 775.84 ns]
parse_bytes/unicode/corpus
                        time:   [1.7091 ms 1.7343 ms 1.7633 ms]
parse_bytes/ascii/corpus
                        time:   [1.0412 ms 1.0510 ms 1.0627 ms]

split/words/empty time:   [17.115 ns 17.232 ns 17.360 ns]
split/words/no_tokens
                        time:   [75.840 ns 76.951 ns 78.219 ns]
split/words/single_token
                        time:   [43.453 ns 43.592 ns 43.745 ns]
split/words/sherlock
                        time:   [1.8230 us 1.8421 us 1.8618 us]
split/words/code  time:   [1.3854 us 1.3997 us 1.4171 us]
split/words/corpus
                        time:   [2.4438 ms 2.4521 ms 2.4614 ms]
`
parse_bytes+split/unicode/empty
                        time:   [92.710 ns 94.199 ns 95.795 ns]
parse_bytes+split/ascii/empty
                        time:   [74.218 ns 75.321 ns 76.539 ns]
parse_bytes+split/unicode/no_tokens
                        time:   [153.64 ns 154.44 ns 155.35 ns]
parse_bytes+split/ascii/no_tokens
                        time:   [85.234 ns 85.434 ns 85.638 ns]
parse_bytes+split/unicode/single_token
                        time:   [225.50 ns 227.78 ns 230.12 ns]
parse_bytes+split/ascii/single_token
                        time:   [120.57 ns 122.55 ns 124.75 ns]
parse_bytes+split/unicode/sherlock
                        time:   [7.9711 us 8.0622 us 8.1804 us]
parse_bytes+split/ascii/sherlock
                        time:   [3.1448 us 3.1599 us 3.1780 us]
parse_bytes+split/unicode/code
                        time:   [5.4211 us 5.4365 us 5.4533 us]
parse_bytes+split/ascii/code
                        time:   [2.2947 us 2.3546 us 2.4133 us]
parse_bytes+split/unicode/corpus
                        time:   [7.6169 ms 7.7082 ms 7.8051 ms]
parse_bytes+split/ascii/corpus
                        time:   [3.5799 ms 3.6196 ms 3.6623 ms]
```

Fixes #225